feat(network-activity-plugin): show headers in SSE requests
