+++ packages/network-activity-plugin/src/react-native/http/network-inspector.ts
@@ -1,33 +1,9 @@
 import { HttpMethod, NetworkActivityDevToolsClient } from '../../shared/client';
-import { getHttpHeaderValue } from '../../ui/utils/getHttpHeaderValue';
+import { getContentType } from '../utils';
 import { getNetworkRequestsRegistry } from './network-requests-registry';
 import { XHRInterceptor } from './xhr-interceptor';
 
 const networkRequestsRegistry = getNetworkRequestsRegistry();
-
-const getContentType = (request: XMLHttpRequest): string => {
-  const responseHeaders = request.responseHeaders;
-  const responseType = request.responseType;
-
-  const contentType = getHttpHeaderValue(responseHeaders || {}, 'content-type');
-
-  if (contentType) {
-    return contentType.split(';')[0].trim();
-  }
-
-  switch (responseType) {
-    case 'arraybuffer':
-    case 'blob':
-      return 'application/octet-stream';
-    case 'text':
-    case '':
-      return 'text/plain';
-    case 'json':
-      return 'application/json';
-    case 'document':
-      return 'text/html';
-  }
-};
 
 const getResponseSize = (request: XMLHttpRequest): number => {
   if (typeof request.response === 'object') {
